wip fix dropping
