+++ src/sdam/description/topology/test/sdam.rs
@@ -578,7 +578,7 @@ async fn topology_closed_event_last() {
     drop(client);
 
     subscriber
-        .wait_for_event(Duration::from_millis(500), |event| {
+        .wait_for_event(Duration::from_millis(5000), |event| {
             matches!(event, Event::Sdam(SdamEvent::TopologyClosed(_)))
         })
         .await
